only add to chain if building on tip
if not building on tip we shouldn't add to chain.

Situation would happen if we are reindexing blocks and we are < maxtipage*2 so we won't hear about nevm block data (which means we only add to mappings). In this case we will end up skipping over blocks and when it comes to >= maxtipage*2 and we get nevm block data we will not be in sequential order. Instead we rely on Geth sync to fetch blocks and we only setup mappings (not insert into chain even if we have the block data)