disconnect only if the chain is synced to tip
disconnecting the head should only happen if the chain actually matches the tip otherwise currentNEVMMappingHash == currentHash will fail. Syscoin should be in sync and so mappings should be correct however geth may not have had time to sync the blocks yet so this would ensure only to act on the chain head if the geth node was synced.